refactor(context/execute): extract governance_position helper (increment 3)
Continues the execute-handler split. Moves compute_governance_position_for_context
into `governance_position.rs` — behavior-preserving, pure code movement.

`pub(super)`, re-imported into the execute flow (one caller in internal_execute).
Also drops two now-orphaned top-level imports (MetaRepository — still imported
locally by the test module; entities::StorageType). CachedOrBlob stays in mod.rs
(it belongs to the module-loading path, not governance computation). Explicit
imports, no `use super::*`.

execute/mod.rs: ~2234 -> ~2105 LOC.

Validated: context lib tests (107) pass.
